// src/webapp/src/app/model/player247.ts
export class Player247 {


  constructor(siteId: string, name: string, position: string, height: string, weight: string, homeTown: string, highSchool: string, year: string, compositeRank: string, rankNational: string, rankPosition: string, rankState: string, stars: string, link: string, college: string) {
    this.siteId = siteId;
    this.name = name;
    this.position = position;
    this.height = height;
    this.weight = weight;
    this.homeTown = homeTown;
    this.highSchool = highSchool;
    this.year = year;
    this.compositeRank = compositeRank;
    this.rankNational = rankNational;
    this.rankPosition = rankPosition;
    this.rankState = rankState;
    this.stars = stars;
    this.link = link;
    this.college = college;
  }

  siteId: string;
  name: string;
  position: string;
  height: string;
  weight: string;
  homeTown: string;
  highSchool: string;
  year: string;
  compositeRank: string;
  rankNational: string;
  rankPosition: string;
  rankState: string;
  stars: string;
  link: string;
  college: string;

}